requirements.txt
Pillow>=2.0.0
requests>=1.1.0
web.py>=0.3.7
